% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{anno_great}
\alias{anno_great}
\title{A profileplyr object that has been annotated with annotateRanges_great()}
\usage{
data(anno_great)
}
\value{
A profileplyr object that has been annotated with annotateRanges_great()/
}
\description{
A profileplyr object that has been annotated with annotateRanges_great(), the profileplyr object from the vignette was used
}
\details{
\itemize{
\item anno_great
}
}
\keyword{datasets}
