% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{annotated_object}
\alias{annotated_object}
\title{A profileplyr object that has been annotated with gene list, GRanges overlap, and genomic annotation types.}
\usage{
data(annotated_object)
}
\value{
A profileplyr object that has been annotated with gene list, GRanges overlap, and genomic annotation types./
}
\description{
profileplyr object that has been annotated with annotateRanges_great(), the profileplyr object from the vignette was used
}
\details{
\itemize{
\item annotated_object
}
}
\keyword{datasets}
