% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{annotateRanges_great}
\alias{annotateRanges_great}
\alias{annotateRanges_great,profileplyr-method}
\title{Annotate profileplyr ranges to genes using local rGREAT}
\usage{
annotateRanges_great(object = "profileplyr", species = "character", ...)

\S4method{annotateRanges_great}{profileplyr}(object = "profileplyr", species = "character", ...)
}
\arguments{
\item{object}{A profileplyr object}

\item{species}{Genome build for the TSS annotation, passed to rGREAT's
\code{tss_source} argument. Accepts "hg19", "hg38", "mm10", "mm9" (or any
TxDb.* package name / other tss_source format rGREAT supports).}

\item{...}{passed to \code{\link[rGREAT]{great}}}
}
\value{
A profileplyr object
}
\description{
The ranges from the deepTools matrix will be annotated to genes using the
local implementation of the GREAT algorithm (\code{\link[rGREAT]{great}}),
which builds extended-TSS regulatory domains from a Bioconductor TxDb and
does not depend on the GREAT web service.
}
\details{
tbd
}
\section{Methods (by class)}{
\itemize{
\item \code{annotateRanges_great(profileplyr)}: Annotate profileplyr ranges to genes using local rGREAT

}}
\examples{
library(SummarizedExperiment)
example <- system.file("extdata", "example_deepTools_MAT", package = "profileplyr") 
object <- import_deepToolsMat(example)
object <- object[1:5, , ] 

# annotate ranges with genes using local GREAT:
annotateRanges_great(object, species = "mm10")

}
