% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soGGi_imports.R
\name{c,ChIPprofile-method}
\alias{c,ChIPprofile-method}
\title{Combine ChIPprofile objects}
\usage{
\S4method{c}{ChIPprofile}(x, ...)
}
\arguments{
\item{x}{A ChIPprofile object}

\item{...}{Additional ChIPprofile objects}
}
\value{
A ChIPprofile object
}
\description{
Combine ChIPprofile objects
}
