% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soGGi_imports.R
\docType{data}
\name{chipExampleBig}
\alias{chipExampleBig}
\title{Example ChIPprofile object}
\usage{
data(chipExampleBig)
}
\value{
A ChIPprofile object
}
\description{
ChIP-seq signal profiles over gene bodies, stored as a ChIPprofile object.
}
