% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soGGi_imports.R
\docType{class}
\name{ChIPprofile-class}
\alias{ChIPprofile-class}
\title{ChIPprofile S4 class}
\description{
An S4 class representing ChIP-seq profiles over genomic intervals.
Extends \code{RangedSummarizedExperiment} with a \code{params} slot
storing the parameters used to generate the profiles.
}
\section{Slots}{

\describe{
\item{\code{params}}{A list of parameters used when generating the profile.}
}}

