% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export.profileplyr}
\alias{export,profileplyr,character-method}
\alias{export_deepToolsMat}
\alias{import_deepToolsMat}
\title{Export profileplyr object}
\usage{
export.profileplyr(object, con)

\S4method{export}{profileplyr,character}(object = "profileplyr",
  con = "character")

export_deepToolsMat(object, con, decreasing = FALSE)

import_deepToolsMat(deeptools_matrix)
}
\arguments{
\item{object}{A profileplyr object}

\item{con}{Connection to write to.}

\item{decreasing}{If metadata(object)$mcolToOrderBy has been set and not NULL, then the ranges will be ordered by the column indicated in this slot of the metadata. By default, the order will be increasing for the factor or numeric value. For decreasing order, choose decreasing = TRUE.}

\item{deeptools_matrix}{Path to deepTools matrix}
}
\value{
The path to deepTools matrix file

A profileplyr object
}
\description{
Export profileplyr object to deepTools matrix file

Export and Import files
}
\details{
A profileplyr object
}
\examples{

example <- system.file("extdata", "example_deepTools_MAT", package = "profileplyr") 
object <- import_deepToolsMat(example) 
export_deepToolsMat(object,"ATAC_Example.MAT")
}
