% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{gene_list_dataframe}
\alias{gene_list_dataframe}
\title{Dataframe of top differentially expressed genes from hindbrain versus liver as measured by RNA-seq}
\usage{
data(gene_list_dataframe)
}
\value{
A dataframe of top differentially expressed genes from hindbrain versus liver as measured by RNA-seq/
}
\description{
This dataset contains a dataframe of the top differentially expressed genes in the hindbrain versus liver as measured by RNA-seq (both genes that go up and those that go down). The gene names are the rownames, and the first column is the 'stat' column from DESeq2. Data was downloaded from ENCODE.
}
\details{
\itemize{
\item gene_list_dataframe
}
}
\keyword{datasets}
