% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soGGi_imports.R
\docType{data}
\name{ik_Profiles}
\alias{ik_Profiles}
\title{Example ChIPprofile object (Ikaros profiles)}
\usage{
data(ik_Profiles)
}
\value{
A ChIPprofile object
}
\description{
ChIP-seq signal profiles over peak regions from two Ikaros antibodies,
stored as a ChIPprofile object.
}
