% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{inherit_group_function}
\alias{inherit_group_function}
\title{Redundant code for inheriting grouping wrapped into subsetbyRangeOverlap() or subsetbyGeneListOverlap() functions}
\usage{
inherit_group_function(object, rowGroupsInUse_input, type, separateDuplicated)
}
\arguments{
\item{object}{A profileplyr object}

\item{rowGroupsInUse_input}{the inherited rowGroupsInUse}

\item{type}{Either "GR" for subsetbyRangeOverlap() function or "GL" for subsetbyGeneListOverlap() function}

\item{separateDuplicated}{A logical argument, if TRUE then regions that overlap multiple inputs to 'GRanges' argument will be separated and made into their own group. All possible combinations of region overlaps will be tested, so it is not recommended to have more than 3 groups if this option is TRUE. If FALSE, then regions that overlap each individual 'GRanges' input will be in the output, and if one region overlaps multiple 'GRanges' inputs, then it will be duplicated in the output and will show up in the section for each group.}
}
\value{
A profileplyr object
}
\description{
Redundant code for inheriting grouping wrapped into subsetbyRangeOverlap() or subsetbyGeneListOverlap() functions
}
\details{
tbd
}
