% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{K27ac_GRlist_hind_liver_top5000}
\alias{K27ac_GRlist_hind_liver_top5000}
\title{GRangesList of the top 5000 H3K27ac peaks from hindbrain and liver downloaded from ENCODE}
\usage{
data(K27ac_GRlist_hind_liver_top5000)
}
\value{
A GRangesList of the top 5000 H3K27ac peaks from hindbrain and liver downloaded from ENCODE/
}
\description{
This dataset contains a GRangesList of the H3K27ac peaks in either the hindbrain or the liver with the highest signal. Data was downloaded from ENCODE.
}
\details{
\itemize{
\item K27ac_GRlist_hind_liver_top5000
}
}
\keyword{datasets}
