% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{params}
\alias{params}
\title{Retrieve and set parameters in profileplyr object}
\usage{
params(object)
}
\arguments{
\item{object}{A profileplyr object}
}
\value{
A list containing parameters for profileplyr object.
}
\description{
Retrieve and set parameters in profileplyr object
}
\examples{
example <- system.file("extdata", "example_deepTools_MAT", package = "profileplyr") 
object <- import_deepToolsMat(example) 
params(object)
}
