% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{profileplyr-class}
\alias{profileplyr-class}
\alias{c,profileplyr-method}
\alias{[,profileplyr,ANY,ANY,ANY-method}
\title{Join, subset and manipulate ChIPprofile objects}
\usage{
\S4method{c}{profileplyr}(x, ...)

\S4method{[}{profileplyr,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{x}{profileplyr object}

\item{...}{Additional arguments.}

\item{i}{An integer or character scalar indicating ranges of profileplyr object to return}

\item{j}{An integer or character scalar indicating columns of profileplyr object to return or a An integer or character scalar indicating which profileplyr object samples to return}

\item{k}{An integer or character scalar indicating samples of profileplyr object to return.}

\item{drop}{A logical whether to drop empty samples}
}
\value{
A profileplyr object
}
\description{
Join, subset and manipulate ChIPprofile objects
}
