% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soGGi_imports.R
\name{pwmToCoverage}
\alias{pwmToCoverage}
\title{PWM hits as an RleList}
\usage{
pwmToCoverage(
  pwm,
  genome,
  min = "70\%",
  removeRand = FALSE,
  chrsOfInterest = NULL
)
}
\arguments{
\item{pwm}{A PWM matrix object.}

\item{genome}{A BSgenome object.}

\item{min}{PWM score cutoff as a percentage of maximum score (default "70\%").}

\item{removeRand}{Logical; remove contigs with "random" in their name.}

\item{chrsOfInterest}{Character vector of chromosomes to include.}
}
\value{
An RleList of motif density per base pair.
}
\description{
Converts a PWM matrix to an RleList of motif hit coverage, suitable for
use as input to \code{regionPlot()}.
}
