% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{subsetbyGeneListOverlap}
\alias{subsetbyGeneListOverlap}
\title{Subset ranges based on overlap with lists of Gene sets}
\usage{
subsetbyGeneListOverlap(
  object,
  group,
  include_nonoverlapping = FALSE,
  separateDuplicated = TRUE,
  inherit_groups = FALSE
)
}
\arguments{
\item{object}{A profileplyr object}

\item{group}{The regions by which to subset the deepTools matrix. This must be either a single GRanges object, or a GRangesList. Combinations of bed file paths and GRanges objects are not accepted, Import BED files as GRanges with rtracklayer import.bed() function.}

\item{include_nonoverlapping}{A logical argument, if FALSE the regions from the original deepTools matrix that do not overlap with the  user defined regions will be left out of the returned profileplyr object.}

\item{separateDuplicated}{A logical argument, if TRUE then regions that overlap multiple inputs to 'GRanges' argument will be separated and made into their own group. All possible combinations of region overlaps will be tested, so it is not recommended to have more than 3 groups if this option is TRUE. If FALSE, then regions that overlap each individual 'GRanges' input will be in the output, and if one region overlaps multiple 'GRanges' inputs, then it will be duplicated in the output and will show up in the section for each group.}

\item{inherit_groups}{A logical whether that groups the exist in the profileplyr object in the 'object' argument should be included in the default grouping scheme for the output object of this function. The default is TRUE. If false, only the gene list overlap annotation will be used for heatmap grouping.}
}
\value{
A profileplyr object
}
\description{
The ranges from the deepTools matrix will be subset based on whether they overlap with user defined gene sets
}
\details{
tbd
}
\examples{
# see the groupby function within profileplyr for examples

}
