% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{annotateRanges_great}
\alias{annotateRanges_great}
\alias{annotateRanges_great,profileplyr-method}
\alias{annotateRanges}
\alias{annotateRanges,profileplyr-method}
\title{Annotate profileplyr ranges to genes using rGREAT or ChIPseeker}
\usage{
annotateRanges_great(object = "profileplyr", species = "character",
  ...)

\S4method{annotateRanges_great}{profileplyr}(object = "profileplyr",
  species = "character", ...)

annotateRanges(object = "profileplyr", annotation_subset = "character",
  TxDb, tssRegion = "numeric", changeGroupToAnnotation = "logical",
  heatmap_grouping = "character", ...)

\S4method{annotateRanges}{profileplyr}(object = "profileplyr",
  annotation_subset = NULL, TxDb, tssRegion = c(-3000, 3000),
  changeGroupToAnnotation = FALSE, heatmap_grouping = "group", ...)
}
\arguments{
\item{object}{A profileplyr object}

\item{species}{GREAT accepts "hg19", "mm10", "mm9", "danRer7" (zebrafish)}

\item{...}{pass to \code{\link[rGREAT]{submitGreatJob}} or \code{\link[ChIPseeker]{annotatePeak}}}

\item{annotation_subset}{If specific annotations (from ChIPseeker package) are desired, specify them here in a character vector. Can be one or any combination of "Promoter", "Exon", "Intron", "Downstream", "Distal Intergenic", "3p UTR", or "5p UTR". This argument is optional and all annotation types will be included if argument is left out.}

\item{TxDb}{TxDb object, a character sting that is a path to a GTF file, or character string indicating genome if one of the following - "hg19", "hg38", "mm9", "mm10".}

\item{tssRegion}{This needs to be a vector of two numbers that will define promoter regions. The first number must be negative, while the second number must be positive. Default values are  c(-3000, 3000) - SHOULD WE CHANGE THIS, SEEMS BIG!)}

\item{changeGroupToAnnotation}{If the grouping should be changed to the annotations (typically when the ranges will be exported for visualization based on this annotation), this should be TRUE. The default if FALSE, which will keep the grouping that existed before annotating the object. This is typical if the output will be used for finding overlaps with gene lists in the 'groupBy' function.}

\item{heatmap_grouping}{Only relevant if 'keepAnnotationAsGroup' is set to TRUE. This argument needs to be either "group", or "annotation". This will determine how the ranges are grouped in the resulting object. Default is heatmap_grouping = "Group". If there are no groups in the deepTools matrix that was used in the function, this argument is unnecessary}
}
\value{
A profileplyr object

A profileplyr object
}
\description{
The ranges from the deepTools matrix will be subset based on whether they overlap with specified annotated regions related to a user defined gene list.

The ranges from the deepTools matrix will be subset based on whether they overlap with specified annotated regions (using ChIPseeker) throughout the genome
}
\details{
tbd

tbd
}
\section{Methods (by class)}{
\itemize{
\item \code{profileplyr}: Annotate profileplyr ranges to genes using rGREAT or ChIPseeker

\item \code{profileplyr}: Annotate profileplyr ranges to genes using rGREAT or ChIPseeker
}}

\examples{

library(SummarizedExperiment)
example <- system.file("extdata", "example_deepTools_MAT", package = "profileplyr") 
object <- import_deepToolsMat(example)
object <- object[1:2, , ] 

# annotate ranges with genes using ChIPseeker 
# (NOTE: can choose subset of annotations with 'annotation_subset' argument)

annotateRanges(object, TxDb = "mm10")

# annotate ranges with genes using GREAT with follwoing command (not run):
# annotateRanges_great(object, species = "mm10")



}
