% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{as_profileplyr}
\alias{as_profileplyr}
\title{Import ChIPprofile object to profileplyr Import}
\usage{
as_profileplyr(chipProfile, names = NULL)
}
\arguments{
\item{chipProfile}{A ChIPprofile object as created by soGGi regionPlot() function.}

\item{names}{Column to select row IDs/names from ChIPprofile mcols.}
}
\value{
A profileplyr object
}
\description{
Working with soGGi ChIPprofile objects in profileplyr.
}
\examples{

library(soGGi)
data("ik_Profiles")
proplyr <- as_profileplyr(ik_Profiles,names="ID")
export_deepToolsMat(proplyr,con="ik_Profiles.MAT")
}
