% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{BamBigwig_to_chipProfile}
\alias{BamBigwig_to_chipProfile}
\title{BamBigwig_to_chipProfile}
\usage{
BamBigwig_to_chipProfile(signalFiles, testRanges,
  testRanges_names = NULL, format, ...)
}
\arguments{
\item{signalFiles}{paths to either BAM files or bigwig files. More than one path can be in this character vector, but all paths in one function call must point to be either all BAM files or all bigWig files, not a combination of the two.}

\item{testRanges}{Either a character vector with paths to BED files.}

\item{testRanges_names}{A character vector with the desired names of the regions corresponding to each element of the 'testRanges' argument. If this is left NULL, the name of each BED file will be used.}

\item{format}{character vector of "bam", "bigwig", "RleList" or "PWM"}

\item{...}{pass to regionPlot() within the soGGi package}
}
\value{
A profileplyr object
}
\description{
Generate a soGGi ChIPprofile object with multiple BAM/bigWig files or multiple BED files as the input
}
\examples{
signalFiles <- c(system.file("extdata",
                             "Sorted_Hindbrain_day_12_1_filtered.bam",
                              package = "profileplyr"))
require(Rsamtools)
for (i in seq_along(signalFiles)){
 indexBam(signalFiles[i])
}
testRanges <- system.file("extdata", 
                          "newranges_small.bed", 
                          package = "profileplyr")
BamBigwig_to_chipProfile(signalFiles, 
                         testRanges, 
                         format = "bam",
                         paired=FALSE,
                         style="percentOfRegion",
                         )
}
