% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{gene_list_character}
\alias{gene_list_character}
\title{Character vector of the top  of differenetial accessible regions from hindbrain versus liver comparison.}
\usage{
data(gene_list_character)
}
\value{
A character vector of the top differenetial accessible regions from hindbrain versus liver comparison.
}
\description{
This dataset contains a character vector of the top differenetial accessible regions from comparison of ATAC-seq signal for hindbrain versus liver
}
\details{
\itemize{
\item gene_list_character
}
}
\keyword{datasets}
