% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{gene_list_dataframe}
\alias{gene_list_dataframe}
\title{Dataframe of differenetial accessible regions from hindbrain versus liver comparison.}
\usage{
data(gene_list_dataframe)
}
\value{
A dataframe of differential accessible regions of hindbrain versus liver in a dataframe/
}
\description{
This dataset contains a data frame of the differential accessible regions from comparison of ATAC-seq signal for hindbrain versus liver
}
\details{
\itemize{
\item gene_list_dataframe
}
}
\keyword{datasets}
