% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{orderBy}
\alias{orderBy}
\alias{orderBy.profileplyr}
\alias{orderBy,profileplyr-method}
\title{orderBy}
\usage{
orderBy.profileplyr(object, column)

orderBy(object = "profileplyr", column = "character")

\S4method{orderBy}{profileplyr}(object = "profileplyr",
  column = "character")
}
\arguments{
\item{object}{A profileplyr object}

\item{column}{Which column of mcols(proplyrObject) should be used for ordering the ranges}
}
\value{
A profileplyr object
}
\description{
choose the column by which to order the ranges by within each group
}
\details{
Takes a profileplyr object and orders the rows based on a user defined metadata column of rowRanges
}
\examples{
example <- system.file("extdata", "example_deepTools_MAT", package = "profileplyr") 
object <- import_deepToolsMat(example) 

library(SummarizedExperiment)
cluster <- clusterRanges(object, fun = rowMeans, cutree_rows = 3)
cluster_order <- orderBy(cluster, column = "hierarchical_order")
metadata(cluster_order)$mcolToOrderBy

}
